# backend/adapters/http/swagger.yaml
openapi: 3.0.3
info:
  title: BloodConnect API
  description: API for managing blood donation requests and donor notifications.
  version: 1.2.0

servers:
  - url: http://localhost:8080
    description: Local development server

paths:
  /users/signup:
    post:
      summary: Register a new user
      tags: [Users]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignupRequest'
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IDResponse'
        '400':
          $ref: '#/components/responses/BadRequest'

  /users/login:
    post:
      summary: Authenticate a user
      tags: [Users]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Successfully authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

  /users/me:
    get:
      summary: Get current authenticated user
      tags: [Users]
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Current user profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'

  /users/me/health:
    put:
      summary: Update health information for the current user
      tags: [Users]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateHealthRequest'
      responses:
        '204':
          description: Health info updated
        '401':
          $ref: '#/components/responses/Unauthorized'

  /users/me/locations:
    post:
      summary: Add a new preferred donation location
      tags: [Users]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddLocationRequest'
      responses:
        '204':
          description: Location added
        '401':
          $ref: '#/components/responses/Unauthorized'

  /users/me/locations/{h3hex}:
    delete:
      summary: Remove a preferred donation location
      tags: [Users]
      security:
        - BearerAuth: []
      parameters:
        - name: h3hex
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Location removed
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'

  /notifications:
    get:
      summary: Get notifications for the current user (Paginated)
      tags: [Notifications]
      security:
        - BearerAuth: []
      parameters:
        - name: last_notification_id
          in: query
          description: ULID of the last notification from the previous page
          schema:
            type: string
      responses:
        '200':
          description: List of notifications
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationsResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

  /requests/{id}/respond:
    post:
      summary: Respond to a donation request as the current user
      tags: [Requests]
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RespondRequest'
      responses:
        '204':
          description: Response recorded
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  /requests/{id}/cancel:
    post:
      summary: Cancel a donation request owned by the current user
      tags: [Requests]
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Request cancelled
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  /requests:
    get:
      summary: List blood donation requests with filters
      tags: [Requests]
      security:
        - BearerAuth: []
      parameters:
        - name: blood_type
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: location_hex
          in: query
          schema:
            type: string
        - name: last_request_id
          in: query
          description: ULID of the last request from the previous page
          schema:
            type: string
      responses:
        '200':
          description: Paginated list of requests
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListRequestsResponse'
    post:
      summary: Create a new blood donation request
      tags: [Requests]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRequestRequest'
      responses:
        '201':
          description: Request created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IDResponse'
        '400':
          $ref: '#/components/responses/BadRequest'

  /requests/{id}:
    get:
      summary: Get details of a specific donation request
      tags: [Requests]
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Detailed request info
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExtendedDonationRequest'
        '404':
          $ref: '#/components/responses/NotFound'

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    IDResponse:
      type: object
      properties:
        id:
          type: string

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        details:
          type: object

    User:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        email: { type: string }
        phone: { type: string }
        health:
          type: array
          items:
            $ref: '#/components/schemas/UserHealth'
        locations:
          type: array
          items:
            $ref: '#/components/schemas/UserLocation'
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }

    SignupRequest:
      type: object
      required: [name, email, password, phone, blood_type, lat, lng]
      properties:
        name: { type: string }
        email: { type: string }
        password: { type: string }
        phone: { type: string }
        blood_type: { type: string }
        lat: { type: number }
        lng: { type: number }

    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string }
        password: { type: string }

    LoginResponse:
      type: object
      properties:
        id: { type: string }
        token: { type: string }

    UserHealth:
      type: object
      properties:
        info_type: { type: string }
        details: { type: string }

    UserLocation:
      type: object
      properties:
        lat: { type: number }
        lng: { type: number }
        h3_hex: { type: string }

    AddLocationRequest:
      type: object
      required: [lat, lng]
      properties:
        lat: { type: number }
        lng: { type: number }

    CreateRequestRequest:
      type: object
      required: [location_lat, location_lng, bag_count, required_by_date, blood_type, location_name]
      properties:
        location_lat: { type: number }
        location_lng: { type: number }
        bag_count: { type: integer }
        required_by_date: { type: string, format: date-time }
        blood_type: { type: string }
        description: { type: string }
        requester_info: { type: string }
        location_name: { type: string }

    DonationRequest:
      type: object
      properties:
        id: { type: string }
        user_id: { type: string }
        blood_type: { type: string }
        bag_count: { type: integer }
        status: { type: string }
        location_name: { type: string }
        location_hex: { type: string }
        location_lat: { type: number }
        location_lng: { type: number }
        required_by_date: { type: string, format: date-time }
        description: { type: string }
        requester_info: { type: string }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }

    ListRequestsResponse:
      type: object
      properties:
        requests:
          type: array
          items:
            $ref: '#/components/schemas/DonationRequest'
        last_request_id: { type: string }
        page_size: { type: integer }

    ExtendedDonationRequest:
      type: object
      properties:
        request:
          $ref: '#/components/schemas/DonationRequest'
        notified_users:
          type: array
          items:
            $ref: '#/components/schemas/ActionedUser'

    ActionedUser:
      type: object
      properties:
        user_id: { type: string }
        action: { type: string }
        lat: { type: number }
        lng: { type: number }
        h3_hex: { type: string }

    RespondRequest:
      type: object
      required: [action]
      properties:
        action:
          type: string
          enum: [Accepted, Declined, Donated]

    Notification:
      type: object
      properties:
        id: { type: string }
        type: { type: string }
        title: { type: string }
        content: { type: string }
        metadata: { type: object }
        created_at: { type: string, format: date-time }

    NotificationsResponse:
      type: object
      properties:
        notifications:
          type: array
          items:
            $ref: '#/components/schemas/Notification'
        last_notification_id: { type: string }
        page_size: { type: integer }

  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Not Found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'